+++ CHANGELOG.md
@@ -8,10 +8,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release template] - 1971-11-03
 
 ### Added
--
 
 ### Changed
 -
+
+### Deprecated
+-
+
+### Removed
+-
+
+### Fixed
+-
+
+### Security
+-
+
+## [Upcoming release template] - 1971-11-03
+
+### Added
+
+### Changed
+- Add a 'by search' variation of the local authority component
 
 ### Deprecated
 -