+++ CHANGELOG.md
@@ -8,9 +8,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release template] - 1971-11-03
 
 ### Added
+-
+
+### Changed
+-
+
+### Deprecated
+-
+
+### Removed
+-
+
+### Fixed
+-
+
+### Security
+-
+
+## [Upcoming release template] - 1971-11-03
+
+### Added
 - Add Form/Tooltip component.
 
 ### Changed
+- Add a 'by search' variation of the local authority component
 - InputField, Radio, Checkbox, Dropdown, Textarea components.
 
 ### Deprecated